fix(config): Move config to app store
