--- src/modules/settings/engages/graphql/queries.ts
@@ -1,13 +1,0 @@
-const engagesConfigDetail = `
-  query engagesConfigDetail {
-    engagesConfigDetail {
-      accessKeyId
-      secretAccessKey
-      region
-    }
-  }
-`;
-
-export default {
-  engagesConfigDetail
-};
