--- src/modules/settings/engages/types.ts
@@ -1,26 +1,0 @@
-export interface IEngageConfig {
-  accessKeyId: string;
-  secretAccessKey: string;
-  region: string;
-}
-
-// query types
-export type EngageConfigQueryResponse = {
-  engagesConfigDetail: IEngageConfig;
-  loading: boolean;
-  refetch: () => void;
-};
-
-// mutation types
-
-export type EngagesConfigSaveMutationResponse = {
-  engagesConfigSave: (
-    params: {
-      variables: {
-        accessKeyId: string;
-        secretAccessKey: string;
-        region: string;
-      };
-    }
-  ) => Promise<any>;
-};
