+++ src/modules/settings/integrations/graphql/queries.ts
@@ -112,6 +112,16 @@ const messengerAppsCount = `
   }
 `;
 
+const engagesConfigDetail = `
+  query engagesConfigDetail {
+    engagesConfigDetail {
+      accessKeyId
+      secretAccessKey
+      region
+    }
+  }
+`;
+
 const fetchApi = `
   query integrationsFetchApi($path: String!, $params: JSON!) {
     integrationsFetchApi(path: $path, params: $params)
@@ -125,6 +135,7 @@ export default {
   integrationDetail,
   integrationTotalCount,
   integrations,
+  engagesConfigDetail,
   messengerApps,
   messengerAppsCount
 };
