+++ src/modules/settings/integrations/types.ts
@@ -153,6 +153,30 @@ export type IntegrationsCountQueryResponse = {
   loading: boolean;
 };
 
+export interface IEngageConfig {
+  accessKeyId: string;
+  secretAccessKey: string;
+  region: string;
+}
+
+export type EngageConfigQueryResponse = {
+  engagesConfigDetail: IEngageConfig;
+  loading: boolean;
+  refetch: () => void;
+};
+
+export type EngagesConfigSaveMutationResponse = {
+  engagesConfigSave: (
+    params: {
+      variables: {
+        accessKeyId: string;
+        secretAccessKey: string;
+        region: string;
+      };
+    }
+  ) => Promise<any>;
+};
+
 export interface IGmailAttachment {
   filename?: string;
   mimeType?: string;
